requirements.txt
action==1.4.4
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
mixins==0.1.4
pkg-resources==0.0.0
pytz==2021.1
routers==0.10.1
sqlparse==0.4.1
